fixed user follows and unfollows

+++ userfollow.py
@@ -5,10 +5,19 @@ import usersearch
 def follow_user(username):
     quit = False
     while(not quit):
-        word = input("What is the email of the user that you would like to follow? ")
+        word = input("What is the username of the user that you would like to follow? ")
         # TODO: SQL NEEDS TO CHECK IF THE USER IS ALREADY BEING FOLLOWED
-        dbaccess.execute_start("INSERT INTO userfollowsuser (username, follows) \
-                VALUES ('"+ username +"', '"+ str(word) +"')")
+        lst = dbaccess.execute_query("SELECT username from users WHERE username = '%s'"%(word))
+        if len(lst) != 0: 
+            lst = dbaccess.execute_query("SELECT username from userfollowsuser WHERE username = '%s' AND follows = '%s'" %(username, word))
+            if len(lst) == 0:
+                dbaccess.execute_start("INSERT INTO userfollowsuser (username, follows) \
+                        VALUES ('"+ username +"', '"+ str(word) +"')")
+                print("User %s has been followed." %(word))
+            else: 
+                print("User %s is already followed!" % (word))
+        else: 
+            print("User %s doesn't exist!" % (word))
         c = input("Would you like to follow another user? ")
         if c.upper()[0] == 'N':
             quit = True
@@ -16,17 +25,21 @@ def follow_user(username):
 def unfollow_user(username):
     quit = False
     while(not quit):
-        word = input("What is the email of the user that you would like to unfollow? ")
-        dbaccess.execute_start("DELETE FROM userfollowsuser \
+        word = input("What is the username of the user that you would like to unfollow? ")
+        lst = dbaccess.execute_query("SELECT username FROM userfollowsuser WHERE username = '%s' AND follows = '%s'" %(username, word))
+        if len(lst) != 0:
+            dbaccess.execute_start("DELETE FROM userfollowsuser \
                 WHERE username = '"+ username +"' AND follows = '"+ str(word) +"'")
+        else: 
+            print("You do not follow this user.")
         c = input("Would you like to unfollow another user? ")
         if c.upper()[0] == 'N':
             quit = True
 
 def follow_screen(username):
     print("What would you like to do?")
-    print("1. follow a user")
-    print("2. unfollow a user")
+    print("1. Follow another user")
+    print("2. Unfollow another user")
     num = input("Enter your selection here: [1, 2] ")
     valid = False
     while not valid:
@@ -40,5 +53,6 @@ def follow_screen(username):
             num = input("Incorrect value. Please try again: [1, 2] ")
 
 if __name__ == '__main__':
-    username = useraccess.login()
+    #username = useraccess.login()
+    username = "hannakoh"
     follow_screen(username)
